+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixes
 
 - Fix "Tag not resolved for Function value" error when serializing objects containing function values in YAML output; functions are now rendered as `[Function]`
+- Fix missing summary output when test produces large amounts of output; stdout is now drained before the process exits
 
 ## [3.9.7] 2025-11-22